MAPR-18949: Date/Time specification update.
With this commit, the behavior of DATE and TIME types has changed to
become timezone independent. The date value of java.sql.Date and the
time value of java.sql.Time returned from Document will be same in all
timezone and will be equal to the one which were set originally.